Add files via upload
This C++ console game was creaed in NetBeans and takes an input for number of players and a sentence to be modified in the game.  The text file link needs to be updated appropriately when the repository is downloaded for the game to work properly. The message is passed from player to player with a chance that the message will be modified.  At the last player, the final message is repeated and the user can see how different the message can be after being passed through several people thus describing the potential for mixed messages in communication.